% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Display Fatal Injuries for a Given Year on a Geographical Map}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{character or numeric, state number}

\item{year}{character or numeric, year}
}
\value{
Map object
}
\description{
Display Fatal Injuries for a Given Year on a Geographical Map
}
\details{
The function reads the data file for the input year (using
  \code{\link{fars_read}}) and displays the fatal injuries using dots on a
  geographical map.
}
