% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read.R
\name{fars_read}
\alias{fars_read}
\title{Read FARS Data}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{character, name of file to read data from.}
}
\value{
A \code{\link{tibble}} of read data
}
\description{
The function reads FARS data from the specified file name or
  URL.
}
\details{
The function relies on the `readr` function \code{\link{read_csv}}
  instead of the base R function. If the file does not exist, the function
  throws an error.
}
\seealso{
\code{\link{make_filename}} to generate filenames that can be used
  by the function.
}
