% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Get Month/Year of Fatal Injuries for Specified Year(s)}
\usage{
fars_read_years(years)
}
\arguments{
\item{years, }{vector or list of years}
}
\value{
List of data frames, where each data frame has two columns for month
  and year
}
\description{
Get Month/Year of Fatal Injuries for Specified Year(s)
}
\details{
The function reads the data file(s) for the input years and returns
  the months of fatal injuries. Warnings will be produced for invalid years.
}
\seealso{
\code{\link{fars_summarize_years}}
}
