% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_years.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Count Fatal Injuries per Month/Year for Specified Year(s)}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years, }{vector or list of years}
}
\value{
A tibble with fatal injury count per month/year
}
\description{
Count Fatal Injuries per Month/Year for Specified Year(s)
}
\details{
The function reads the data file(s) for the input years (using
  \code{\link{fars_read_years}}) and counts the injuries within each month.
}
