% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Create a FARS File Name for a Given Year}
\usage{
make_filename(year)
}
\arguments{
\item{year}{numeric or character, year to be used in the file name generation}
}
\value{
Generated file name as a character
}
\description{
To facilitate reading data from FARS files, this function takes
  a year to generate the corresponding FARS data file name
}
\details{
The function uses \code{\link{sprintf}} to generate the file name,
  after coercing the `year` argument to an integer.
}
\seealso{
\code{\link{fars_read}}
}
